requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
openai
openai-agents
python-docx
pypdf
jinja2